scripts task works as requirements describe
As a developer, I should be able to run the gulp scripts command at the
command line to concatenate, minify, and copy all of the project’s
JavaScript files into an all.min.js file that is then copied to the
dist/scripts folder.

———————————

Used gulp-pump to handle errors and gulp-rename to copy the stream both
as an all.js and an all.min.js file so that later in the Gulp build
that errors can be matched to a non-minified file.